x_bot/requirements.txt
tweepy>=4.14.0
anthropic>=0.40.0
apscheduler>=3.10.0
yfinance>=0.2.30
python-dotenv>=1.0.0
loguru>=0.7.0
pytz>=2024.1
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
